+++ src/pages/HomePage/FilterTextInput.tsx
@@ -1,0 +1,11 @@
+import { TextField } from "@mui/material";
+
+const FilterTextInput = () => {
+  return (
+    <>
+      <TextField />
+    </>
+  );
+};
+
+export default FilterTextInput;