+++ src/utils/dummyCategories.ts
@@ -1,0 +1,21 @@
+import categoryType from "../types/categoryType";
+
+const dummyCategories: categoryType[] = [
+  {
+    id: "1",
+    name: "Travel",
+    colour: "#0000ff",
+  },
+  {
+    id: "2",
+    name: "Work",
+    colour: "#ff00ff",
+  },
+  {
+    id: "3",
+    name: "Expenses",
+    colour: "#00ffff",
+  },
+];
+
+export default dummyCategories;